+++ projects/mk-kit/directives/pipes/index.ts
@@ -1,0 +1,6 @@
+export * from './currency.pipe';
+export * from './relative-time.pipe';
+export * from './file-size.pipe';
+export * from './initials.pipe';
+export * from './truncate.pipe';
+export * from './pluralize.pipe';